have localStorage items on the global enivroment to ensure it displays on the DOM, and provided some functionality that is like the displayInfo function. Change one line of code from the showDelBtn function from wrapContent.remove() to contentContainer.remove()

+++ app.js
@@ -3,7 +3,90 @@ const typeHere = document.querySelector('#typeHere')
 const clickBtn = document.querySelector('#clickBtn');
 
 //this is where localStorage appends the DOM
+if(localStorage.getItem('content')){
+    const getLength = JSON.parse(localStorage.getItem('content'))
+    if(getLength.length === 1){
+        let containedText = document.createElement('p');
+        containedText.innerText = getLength[0].containedText
 
+        const showDelBtn = document.createElement('button');
+        showDelBtn.innerText = 'Delete';
+
+        let showEditBtn = document.createElement('button');
+        showEditBtn.innerText = 'Edit';
+
+        const wrapContent  = document.createElement('div');
+        wrapContent.append(containedText, showEditBtn, showDelBtn)
+
+        const contentContainer = document.createElement('div');
+
+        wrapContent.setAttribute('id', 'container1')
+        contentContainer.setAttribute('id', 'content1')
+        containedText.setAttribute('id', 'text1')
+        showDelBtn.setAttribute('id', 'del1')
+        showEditBtn.setAttribute('id', 'edit1')
+
+        contentContainer.append(wrapContent)
+
+        let updateBtn = document.createElement('button');
+        updateBtn.setAttribute('id', 'update1')
+        updateBtn.innerText = 'Update'
+    
+        let inputEditText = document.createElement('input')
+        inputEditText.setAttribute('value', getLength[0].containedText)
+        inputEditText.setAttribute('id', 'editHere1')
+
+        showEditBtn.addEventListener('click', () => {
+            wrapContent.replaceChild(inputEditText, wrapContent.childNodes[0])
+            wrapContent.replaceChild(updateBtn,wrapContent.childNodes[1])
+        })
+
+        updateBtn.addEventListener('click', () => {
+            let updatedText = inputEditText.value
+            containedText.innerText = updatedText
+            wrapContent.replaceChild(containedText, wrapContent.childNodes[0])
+            wrapContent.replaceChild(showEditBtn,wrapContent.childNodes[1])
+            // memoryObj = [{
+            //     id: 1,
+            //     containedText: getText
+            // }]
+            let changeContent = JSON.parse(localStorage.getItem('content'))
+            console.log(changeContent)
+            console.log(changeContent[0])
+            console.log(changeContent.length)
+            changeContent[0].containedText = updatedText
+            localStorage.setItem('content', JSON.stringify(changeContent))
+        })
+
+        const parentContainer = document.querySelector('#parentContainer');
+        parentContainer.append(contentContainer); 
+
+        showDelBtn.addEventListener('click', () => {
+            // console.log('testing delete button')
+            // console.log(JSON.parse(localStorage.getItem('content')))
+            let getAnswer = confirm('Are you sure you want to delete this specific to do task?')
+    
+            if(getAnswer){
+                contentContainer.remove()
+                let removeElement =JSON.parse(localStorage.getItem('content'))
+                // console.log(removeElement)
+                removeElement.splice(0, 1)
+                // console.log('now: ', removeElement)
+                // console.log(removeElement.length)
+                let placeElementBack = JSON.stringify(removeElement)
+                localStorage.setItem('content', placeElementBack)
+                let getLength = JSON.parse(localStorage.getItem('content'))
+                // console.log(getLength.length)
+                if(getLength.length === 0){
+                    localStorage.clear(); 
+                }
+                // let placeElementBack = JSON.stringify(removeElement)
+                // localStorage.setItem('content', placeElementBack)
+            }
+        })
+    }
+
+}
 
 // When submiting a text
 const displayInfo = () => {
@@ -69,6 +152,9 @@ const displayInfo = () => {
         //     containedText: getText
         // }]
         let changeContent = JSON.parse(localStorage.getItem('content'))
+        console.log(changeContent)
+        console.log(changeContent[0])
+        console.log(changeContent.length)
         changeContent[0].containedText = updatedText
         localStorage.setItem('content', JSON.stringify(changeContent))
         // console.log(containedText.innerText)
@@ -84,7 +170,7 @@ const displayInfo = () => {
         let getAnswer = confirm('Are you sure you want to delete this specific to do task?')
 
         if(getAnswer){
-            wrapContent.remove()
+            contentContainer.remove()
             let removeElement =JSON.parse(localStorage.getItem('content'))
             // console.log(removeElement)
             removeElement.splice(0, 1)
@@ -123,13 +209,16 @@ deleteAllBtn.addEventListener('click', () => {
     // console.log(removeAllChildren.hasChildNodes())
     let getChildElementsLength = removeAllChildren.children.length
     if(localStorage.getItem('content')){
-        localStorage.clear();
-        if(getChildElementsLength > 0){
-            while(removeAllChildren.hasChildNodes()){
-                removeAllChildren.removeChild(removeAllChildren.firstChild);
+        if(confirmChoice){
+            localStorage.clear();
+            if(getChildElementsLength > 0){
+                while(removeAllChildren.hasChildNodes()){
+                    removeAllChildren.removeChild(removeAllChildren.firstChild);
+                }
             }
+            alert('everything has been deleted')
         }
-        alert('everything has been deleted')
+
     }else{
         alert("there's nothing to be deleted")
     }
